+++ Razor/src/Razor/Window.h
@@ -10,12 +10,12 @@ namespace Razor {
 	struct WindowProps
 	{
 		std::string Title;
-		unsigned int Width;
-		unsigned int Height;
+		float Width;
+		float Height;
 
 		WindowProps(const std::string& title = "Razor Engine",
-			unsigned int width = 1280,
-			unsigned int height = 720)
+			float width = 1280.0f,
+			float height = 720.0f)
 			: Title(title), Width(width), Height(height)
 		{
 		}
@@ -30,8 +30,8 @@ namespace Razor {
 
 		virtual void OnUpdate() = 0;
 
-		virtual unsigned int GetWidth() const = 0;
-		virtual unsigned int GetHeight() const = 0;
+		virtual float GetWidth() const = 0;
+		virtual float GetHeight() const = 0;
 
 		virtual void SetEventCallback(const EventCallbackFn& callback) = 0;
 		virtual void SetVSync(bool enabled) = 0;
